ai_engine/requirements.txt
flask==2.3.3
python-dotenv==1.0.0
opencv-python==4.8.0.76
numpy==1.24.3
requests==2.31.0
torch==2.0.1 